refactor: clean up legacy runtime compatibility
